[minor] Noop cleanup for metadata cache
