+++ src/moonlink/src/storage/cache/metadata.rs
@@ -1,10 +1,7 @@
 pub(crate) mod base_cache;
-pub mod cache_config;
-
-pub mod moka_cache;
-
+pub(crate) mod cache_config;
+pub(crate) mod moka_cache;
 #[cfg(test)]
 mod moka_cache_test;
-
 #[cfg(test)]
-pub(crate) mod test_utils;
+mod test_utils;
