+++ src/moonlink/src/storage/cache/metadata/moka_cache.rs
@@ -6,9 +6,6 @@ use crate::storage::cache::metadata::{
 use async_trait::async_trait;
 use moka::future::Cache;
 
-#[cfg(test)]
-mod test_utils;
-
 /// A wrapper around [`moka::future::Cache`] providing async cache operations.
 ///
 /// # Eviction Policy
@@ -25,14 +22,8 @@ mod test_utils;
 /// - **Max size**: limits the number of entries (or total weight if using a custom weigher).
 /// - **Asynchronous operations**: all API methods are `async`.
 ///
-/// # Example
-/// ```rust
-/// let config = MetadataCacheConfig::new(100, Duration::from_secs(60));
-/// let cache: MokaCache<String, String> = MokaCache::new(config);
-/// cache.put("key".to_string(), "value".to_string()).await;
-/// ```
 pub struct MokaCache<K, V> {
-    cache: Cache<K, V>,
+    pub(crate) cache: Cache<K, V>,
 }
 
 #[allow(dead_code)]
@@ -58,30 +49,10 @@ where
     K: std::hash::Hash + Eq + Clone + Send + Sync + 'static,
     V: Clone + Send + Sync + 'static,
 {
-    /// Retrieves a value for the given key.
-    ///
-    /// **Note:** This returns a cloned copy of the value stored in the cache.
-    /// Modifying the returned value does not affect the cached value.
     async fn get(&self, key: &K) -> Option<V> {
         self.cache.get(key).await
     }
 
-    /// Inserts a key-value pair into the cache.
-    ///
-    /// - If the key already exists:
-    ///   - The old value will be **overwritten** with the new one.
-    ///   - The entry's **expiration time (TTL)** will be refreshed (reset from now).
-    ///
-    /// - If the key does not exist:
-    ///   - A new entry will be inserted with the configured TTL.
-    ///
-    /// # Example
-    /// ```
-    /// cache.put("user:1".to_string(), "Alice".to_string()).await;
-    /// cache.put("user:1".to_string(), "Bob".to_string()).await;
-    ///
-    /// // "Alice" is replaced by "Bob", and TTL is refreshed.
-    /// ```
     async fn put(&self, key: K, value: V) {
         self.cache.insert(key, value).await;
     }
@@ -90,34 +61,7 @@ where
         self.cache.invalidate_all();
     }
 
-    /// Removes the entry for the specified `key` from the cache.
-    ///
-    /// # Behavior
-    /// - If the key **exists**:
-    ///   The entry is removed and the previous value is returned as `Some(V)`.
-    ///
-    /// - If the key **does not exist**:
-    ///   Returns `None`.
-    ///   No panic, no error, no log - the operation silently passes.
-    ///
-    /// # Example
-    /// ```
-    /// use your_crate::MokaCache;
-    ///
-    /// # async fn example() {
-    ///     let cache = MokaCache::new();
-    ///     cache.insert("a", 1).await;
-    ///
-    ///     assert_eq!(cache.evict(&"a").await, Some(1)); // removed value
-    ///     assert_eq!(cache.evict(&"a").await, None);    // already gone, silently ignored
-    /// # }
-    /// ``
-    async fn evict(&self, key: &K) -> Option<V> {
+    async fn remove(&self, key: &K) -> Option<V> {
         self.cache.remove(key).await
-    }
-
-    async fn len(&self) -> u64 {
-        self.cache.run_pending_tasks().await;
-        self.cache.entry_count()
     }
 }
